refactor(project-overview): rename project-overview to projects
