+++ pages/projects/_project.vue
